+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -330,7 +330,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTB6K60JRA6Z0WS8753R5G4Y"
+        "01KTJVKQSKHEQGNGTCZ17Y5BQX"
     )
     r"""The unique identifier of the data source"""