+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTB6K60FX3WZR6WQ2QV74ZKN"
+        "01KTJVKQSG1KGTDHRW2H68DSGC"
     )
     r"""The unique identifier of the data source"""