+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTB6K60N4R7W3MKW1TH8MXAA"
+        "01KTJVKQSP7EKB2Z7YAD59WSZ1"
     )
     r"""The unique identifier of the data source"""