+++ packages/orq-rc/RELEASES.md
@@ -8519,3 +8519,13 @@ Based on:
 - [python v4.11.0-rc.1] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.1] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.1 - packages/orq-rc
+
+## 2026-06-08 05:36:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.771.0 (2.893.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.2] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.2] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.2 - packages/orq-rc